Ported l2-regress, l2-tools, mba to new compiler
Checking for java1.7 and java1.8 also
Updated version of Tcl to 8.3.3
Updated version of Tk to 8.3.3
Updated version of itcl to 3.2
Updated version of Tktable to 2.10
Updated version of Combat to 0.8.1
Updated version of Mico to 2.3.13
Checking for iostream and fstream headers
Configure in mpa is using external xerces for now
Using newer version of autoconf, configure files updated

+++ mba/cpp/include/readers/from_l2.h
@@ -47,7 +47,7 @@ class from_l2 : public L2_file_reader {
   size_t line_number;
 
   /// Character index in parsing, with respect to the current line.
-  unsigned index;
+  size_t index;
 
   // Flags to make sure we don't parse the same thing twice.
